Improved the search efficiency by removing resource contention for the view set, and also eliminated a set of concurrent modification exceptions related to rapid changes of the view position set while dynamically recomputing the hierarchy view.
Created a dynamic programming method for isParentOf that significantly improves performance for large or repeated tests. 

Fixed drag and drop issues with the search panels. 

Enabled regular expression search. 

Improved layout and user feedback for search.

Fixed defect in TulpeAdder that was resulting in always returning the latest version on a branch rather than proper results for historic queries (bug was introduced as part of the database optimization).