TCP-2316 Updated BinaryChangeSetWriter added private <P extends I_AmPart> void removeDuplicateParts(List<? extends I_AmVersioned<P>> versionedList), this is called from writeChanges to remove duplicate parts and empty versioned objects. Added test class TestBinaryChangeSetWriterRemoveDuplicateParts.
TCP-2344 Changed id validation so the id in the mapping database is used as this is created from the last export ids. need to fix/remove this logic as we should be using s base lined database. Updated the CopyHierarchyToPath code so the conflict resolution strategy is not used (as wrong as this sounds) this mean it will function as per R1