Artifact artf221129 : Report: Worklist members by status report Tracker: 	User feedback Title: 	Report: Worklist members by status report Description:
We have tried to make a report – Worklist members by status report. When all members in the worklist were in state ‘
Approved for publication status’ the report shows FSN in English only. We need to see the translated Danish term – the
 PT. (Pic28)

We tried to create the same report in another Worklist. Here the members were in status ‘Rejected by TSP reviewer 
status’ and ‘Translated status’. In this case it seems that the PT in English that is shown in the report. We need to
 see the translated term in Danish as well and also the FSN in English as the PT is not unique. (Pic29 will be send in a
 new artifact).