[artf226109] Workflow search is missing concepts.
A few problems discovered with the workflow initiation process. One was that the folder path expectations did not take into account multi-module build projects, so the workflow index folder was being put at the top level project, rather than the project-bdb folder. Fixed this issue by adding a workflow index folder paramater to the change set import mojo. Modified the WfHxLuceneManager to make this variable public so it could be set by the change set import mojo. 

Second issue was that the workflow history did not even write workflow to either the incorrect top level folder, or to the correct project-bdb folder (after the above fix). There was a conditional test that was looking for the workflow runtime directory (which does not exist at this step in the editor assembly process), and if it did not, then no workflow history was written to a lucene index. 

Pom files and archetype need to be modified to have new pom configuration to support workflow history.