FIXED
TC000421 - Validate 'description type' search option

At a lower resolution, the drag and drop functionality of refsets into combo boxes (the two in the left hand side) is 
not working. This is especially apparent when a long string for language refset is present. The string is pushed to the 
left covering the area where the refsets are supposed to be dragged/dropped.