Updated to latest from tags/wb-mmb-1.2-july-id
This is a bit painful, because:

we came from:
/tags/wb-mmb-2.32-trek-no-jini@13446 which came from /branches/trek-no-jini@13399 

But trek-no-jini was abandoned.  So this commit needs all of the patches required to get us over to tags/wb-mmb-1.2-july-id

So we have to backtrack and figure out where we have a common branch, and then not remerge any duplicate commits....

That road goes like this: 

tags/wb-mmb-2.33-trek-no-jini came from /branches/trek-no-jini@13477
branches/trek-no-jini-2.33-id-july came from /tags/wb-mmb-2.33-trek-no-jini@13483
tags/wb-mmb-1.2-july-id came from /branches/trek-no-jini-2.33-id-july@13889

So, there we have it.  To bring our branch up to date, we need the following commit set from branches/trek-no-jini-2.33-id-july

13487,13536-13570,13596,13613-13753,13813-13819,13879-13880

(note, those are the inclusive numbers, not the exclusive-inclusive format that SVN uses in commands)
(when entering these into SVN, subtract 1 from the beginning of all ranges)
(13575 is skipped, as we already have it, same with 13586 and 13606, 13590 skipped because it is fubar - manually merged.)

And this brings us in sync with  tags/wb-mmb-1.2-july-id which is equivalent to branches/trek-no-jini-2.33-id-july@13880